#194: move TransferActivityTracker to transfer (domain) package; test banner iOS predicate
