+++ composeApp/src/androidUnitTest/kotlin/com/tubetoast/tether/network/AndroidTransferLockHolderTest.kt
@@ -2,6 +2,7 @@ package com.tubetoast.tether.network
 
 import android.content.Context
 import com.tubetoast.tether.TetherApp
+import com.tubetoast.tether.transfer.DefaultTransferActivityTracker
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
