+++ composeApp/src/commonMain/kotlin/com/tubetoast/tether/transfer/TransferActivityTracker.kt
@@ -1,4 +1,4 @@
-package com.tubetoast.tether.network
+package com.tubetoast.tether.transfer
 
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.flow.MutableStateFlow
