+++ composeApp/src/commonTest/kotlin/com/tubetoast/tether/presentation/RootComponentTest.kt
@@ -8,7 +8,6 @@ import com.arkivanov.essenty.statekeeper.StateKeeperDispatcher
 import com.tubetoast.tether.config.DeviceNameStore
 import com.tubetoast.tether.config.EphemeralDeviceNamePersistence
 import com.tubetoast.tether.discovery.FakeDeviceDiscovery
-import com.tubetoast.tether.network.NoOpTransferActivityTracker
 import com.tubetoast.tether.peer.PeersRepository
 import com.tubetoast.tether.preferences.FakeFileTransferPreferences
 import com.tubetoast.tether.preferences.FakePeerPreferencesStore
@@ -20,6 +19,7 @@ import com.tubetoast.tether.protocol.Device
 import com.tubetoast.tether.protocol.DeviceType
 import com.tubetoast.tether.transfer.FakeFilePicker
 import com.tubetoast.tether.transfer.FakeFileSource
+import com.tubetoast.tether.transfer.NoOpTransferActivityTracker
 import com.tubetoast.tether.transfer.PeerTransferEngine
 import com.tubetoast.tether.transfer.PeerTransferState
 import com.tubetoast.tether.transfer.PendingFilesRepository
